refactor(osb): separate GitOperationContext from Factory
