+++ src/main/kotlin/io/meshcloud/dockerosb/persistence/GitHandler.kt
@@ -2,6 +2,9 @@ package io.meshcloud.dockerosb.persistence
 
 import java.io.File
 
+/**
+ * Note: consumers should use this only via a [GitOperationContext]
+ */
 interface GitHandler {
 
   fun pullFastForwardOnly()
